send and receive bytes to and from the HAProxy socket
Python 3 distinguishes between Unicode `str`ings and byte strings.
The socket is expecting bytes, and we've dealing with `str`ings
internally so we need to convert in each direction.

This fixes the following problems:

```
Unhandled python exception in read callback: TypeError: a bytes-like object is required, not 'str'
Unhandled python exception in read callback: TypeError: must be str, not bytes
```